Event system
- Add Event System: ApplicationEvent, KeyEvent and MouseEvent.
- Modify log pattern.
- Change data type of MouseButton from uint8_t to uint16_t.

+++ Grafix/src/Grafix/Input/Codes.h
@@ -5,7 +5,7 @@ namespace Grafix
 {
     // Adapted from glfw3.h
 
-    enum class MouseButton : uint8_t
+    enum class MouseButton : uint16_t
     {
         Button1 = 0,
         Button2 = 1,
